Scratch Comparison Between Deep Architectures
We compare five architectures, namely [resnet, alexnet, vgg, squeezenet, densenet] on a image classification dataset. For the first case, we use them as a feature extractor and only train the final layer of the network and for the second case we train these archiectures from scratch, optimizing weights in all the layers. We further plot results from each of these cases.